Sticky Header
Made the header "stick" by making it so that it is always displaying on the top of the screen. This should be changed with javascript later on. I'd like to make it so that thte navbar is see-through a little bit once you scroll down far enough. Also made a divider line. All you ave to do now when you wnat a divider line is do <div class='divider'></div>